sphere texture mapping

+++ src/be/sandervl/raytracer/business/math/MathUtils.java
@@ -2,7 +2,7 @@ package be.sandervl.raytracer.business.math;
 
 public class MathUtils {
 
-    public static float round(int lower, int upper, float n) {
+    public static float round(float lower, float upper, float n) {
         if (n > upper) {
             return upper;
         }
